+++ packages/lumx-react/src/components/dropdown/Dropdown.tsx
@@ -5,11 +5,9 @@ import classNames from 'classnames';
 import { List } from '@lumx/react/components/list/List';
 import { Offset, Placement, Popover } from '@lumx/react/components/popover/Popover';
 import { COMPONENT_PREFIX } from '@lumx/react/constants';
-import { useCallbackOnEscape } from '@lumx/react/hooks/useCallbackOnEscape';
 import { useFocus } from '@lumx/react/hooks/useFocus';
 import { useInfiniteScroll } from '@lumx/react/hooks/useInfiniteScroll';
 import { GenericProps, getRootClassName, handleBasicClasses, isComponent } from '@lumx/react/utils';
-import { ClickAwayProvider } from '@lumx/react/utils/ClickAwayProvider';
 
 /**
  * Defines the props of the component.
@@ -20,7 +18,7 @@ interface DropdownProps extends GenericProps {
     /** Children of the Dropdown. */
     children: React.ReactNode;
     /** Whether a click anywhere out of the Dropdown would close it. */
-    closeOnClick?: boolean;
+    closeOnClickAway?: boolean;
     /** Whether an escape key press would close the Dropdown. */
     closeOnEscape?: boolean;
     /** Whether the dropdown should fit to the anchor width */
@@ -32,7 +30,7 @@ interface DropdownProps extends GenericProps {
     /** Whether the focus should be set on the list when the dropdown is open. */
     shouldFocusOnOpen?: boolean;
     /** Whether the dropdown should be displayed or not. Useful to control the Dropdown from outside the component. */
-    showDropdown: boolean;
+    isOpen: boolean;
     /** The z-axis position. */
     zIndex?: number;
     /** The function to be called when the user clicks away or Escape is pressed */
@@ -62,12 +60,12 @@ const CLASSNAME: string = getRootClassName(COMPONENT_NAME);
  * The default value of props.
  */
 const DEFAULT_PROPS: DefaultPropsType = {
-    closeOnClick: true,
+    closeOnClickAway: true,
     closeOnEscape: true,
     fitToAnchorWidth: true,
     placement: Placement.AUTO_START,
     shouldFocusOnOpen: true,
-    showDropdown: undefined,
+    isOpen: undefined,
 };
 
 /**
@@ -79,13 +77,13 @@ const Dropdown: React.FC<DropdownProps> = ({
     anchorRef,
     children,
     className,
-    closeOnClick = DEFAULT_PROPS.closeOnClick,
+    closeOnClickAway = DEFAULT_PROPS.closeOnClickAway,
     closeOnEscape = DEFAULT_PROPS.closeOnEscape,
     fitToAnchorWidth = DEFAULT_PROPS.fitToAnchorWidth,
     offset,
     placement = DEFAULT_PROPS.placement,
     shouldFocusOnOpen = DEFAULT_PROPS.shouldFocusOnOpen,
-    showDropdown,
+    isOpen,
     zIndex,
     onClose,
     onInfiniteScroll,
@@ -115,22 +113,23 @@ const Dropdown: React.FC<DropdownProps> = ({
         );
     }, [listElementRef, wrapperRef, children, className, props]);
 
-    useCallbackOnEscape(onClose, showDropdown && closeOnEscape);
-
     // Set the focus on the list when the dropdown opens,
     // in order to enable keyboard controls.
     useFocus(listElementRef.current, shouldFocusOnOpen);
 
-    return showDropdown ? (
+    return isOpen ? (
         <Popover
             anchorRef={anchorRef}
             placement={placement}
             offset={offset}
             zIndex={zIndex}
+            fitToAnchorWidth={fitToAnchorWidth}
+            isOpen={isOpen}
+            onClose={onClose}
+            closeOnClickAway={closeOnClickAway}
+            closeOnEscape={closeOnEscape}
         >
-            <ClickAwayProvider callback={closeOnClick && onClose} refs={[wrapperRef, anchorRef]}>
-                {popperElement}
-            </ClickAwayProvider>
+            {popperElement}
         </Popover>
     ) : null;
 };